refactor: extract DirectInput into audio_input_direct module
Keep audio_input_sender.py purely abstract (AudioInputHost Protocol +
AudioInputSender ABC) and move the concrete DirectInput implementation
into its own file. The utils folder now reads as one file per layer:
contract, default implementation, specialized implementation.
