+++ agents-core/vision_agents/core/utils/audio_input_sender.py
@@ -27,27 +27,3 @@ class AudioInputSender(ABC):
 
     @abstractmethod
     async def close(self) -> None: ...
-
-
-class DirectInput(AudioInputSender):
-    """Forward every PCM chunk straight to the host's provider.
-
-    Subclasses may override `send` to add buffering/pacing and call
-    `super().send(...)` to deliver each chunk through the default path.
-    """
-
-    def __init__(self, host: AudioInputHost) -> None:
-        self._host = host
-
-    async def send(self, pcm: PcmData, participant: Participant | None) -> None:
-        if not self._host.connected:
-            return
-        if participant is None:
-            return
-        await self._host.simple_audio_response(pcm, participant)
-
-    def clear(self) -> None:
-        pass
-
-    async def close(self) -> None:
-        pass
